shallow copy fields instead of deep copying
Shallow copying is more efficient and still resolves the original
issue due to which copying was introduced (#17).

Deep copying of a field can fail due to a number of reasons, such as:

1. Fields often indirectly refer to models. A model with relationship
fields uses transient cache attributes which can change during the
copying, resulting in a run-time error.

2. Fields can have direct or indirect references to objects which
cannot be copied, such as HTTP request objects.