Allow escaping of `:` character when processing field attributes
Currently, the `attr` and similar filters simply split the argument into attribute name and attribute value by `:`. This is fine in most cases, however it could sometimes be desirable to be able to insert a `:` inside of the name of the attribute (for example, using [Vue.js](https://vuejs.org/) requires this functionality). For this reason, I made it so that a string like `b:d` gets split into `b` and `d`, but a string like `b::c:d` gets split into `b:c` and `d`.